email_send_backend/requirements.txt
requests
django>=3.2.14
django-cors-headers==4.7.0
djangorestframework==3.16.0
psycopg2-binary==2.9.10